@@ -3,47 +3,60 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
+    "start": "npm run start:dev",
+    "build": "npm run build:dev",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "start:dev": "ng serve --c=dev",
+    "emulator": "firebase emulators:start --import minData --only firestore,auth",
+    "build:dev": "ng build --c=dev",
+    "start:prod": "ng build --c=prod",
+    "build:prod": "ng build --c=prod"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.2",
-    "@angular/cdk": "^10.2.7",
-    "@angular/common": "~13.2.2",
-    "@angular/compiler": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
-    "ng-recaptcha": "^9.0.0",
-    "rxjs": "^7.4.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "~15.2.5",
+    "@angular/cdk": "^15.2.5",
+    "@angular/common": "~15.2.5",
+    "@angular/compiler": "~15.2.5",
+    "@angular/core": "~15.2.5",
+    "@angular/fire": "~7.5.0",
+    "@angular/forms": "~15.2.5",
+    "@angular/localize": "^15.2.5",
+    "@angular/material": "^15.2.5",
+    "@angular/platform-browser": "~15.2.5",
+    "@angular/platform-browser-dynamic": "~15.2.5",
+    "@angular/router": "~15.2.5",
+    "@ngrx/effects": "15.4.0",
+    "@ngrx/entity": "15.4.0",
+    "@ngrx/store": "15.4.0",
+    "@ngrx/store-devtools": "15.4.0",
+    "firebase": "9.19.1",
+    "firebaseui": "^6.0.2",
+    "rxjs": "^7.8.0",
+    "tslib": "^2.5.0",
+    "uuid": "^9.0.0",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.3",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler-cli": "~13.2.2",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
+    "@angular-devkit/build-angular": "^15.2.4",
+    "@angular/cli": "~15.2.4",
+    "@angular/compiler-cli": "^15.0.0",
+    "@types/jasmine": "~4.3.1",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^18.15.11",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~4.6.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "protractor": "^7.0.0",
+    "ts-node": "~10.9.1",
     "tslint": "~6.1.0",
-    "typescript": "~4.5.5"
+    "typescript": "4.8.2"
   }
 }
