@@ -1,77 +1,119 @@
 import { Injectable } from '@angular/core';
 import { Group } from '../models/group-model';
-import { WordManageService } from './word-manage.service';
-import { HttpClient } from '@angular/common/http';
-import { ReplaySubject } from 'rxjs';
-import { environment } from '../../environments/environment';
-import { map } from 'rxjs/operators';
+import { from, Observable, of } from 'rxjs';
+import { concatMap, map } from 'rxjs/operators';
+import { AngularFirestore } from '@angular/fire/compat/firestore';
+import { convertSnaps } from './db-utils';
+import { Word } from '@app/models/word-model';
+
+import { UserService } from '../login/user.service';
+import { AngularFireAuth } from '@angular/fire/compat/auth';
+import { QuerySnapshot, DocumentChange } from '@firebase/firestore-types';
 
 @Injectable({
   providedIn: 'root',
 })
 export class GroupService {
-  BACKEND_URL = environment.apiUrl + '/groups';
-  // GET to get all groups, POST to create a group
-  groupNum = '';
-  // GET to get, PUT to edit, DELETE to delete a single group
-  getGroupUrl = '/groups/' + this.groupNum;
+
 
   constructor(
-    private wordService: WordManageService,
-    private http: HttpClient
-  ) {}
+    private db: AngularFirestore, private userService: UserService, private afAuth: AngularFireAuth
+  ) { }
 
-  private groupsObs = new ReplaySubject<Group[]>(1);
-  private groups: Group[];
 
-  getGroupsFromServer() {
-    return this.http
-      .get<{ message: string; data: any }>(this.BACKEND_URL)
+  createGroup(newGroup: Partial<Group>, groupId?: string): Observable<Partial<Group>> {
+
+    let save$: Observable<any>;
+    if (groupId)
+      save$ = from(this.db.doc(`users/${this.userService.user.uid}/groups/${groupId}`).set(newGroup));
+    else save$ = from(this.db.collection(`users/${this.userService.user.uid}/groups/`).add(newGroup));
+
+    return save$.pipe(
+      map((res) => {
+        return {
+          id: groupId ?? res.id,
+          ...newGroup,
+        };
+      })
+    );
+  }
+
+  updateGroup(groupId: string, changes: Partial<Group>): Observable<any> {
+    return from(this.db.doc(`groups/${groupId}`).update(changes));
+  }
+
+  loadGroups(): Observable<Group[] | any> {
+    console.log("loadGroups()")
+    // return of(["test"]);
+    return this.db
+      .collection(`users/${this.userService.user.uid}/groups`)
+      .get()
+      .pipe(map((results) => convertSnaps<Group>(results)));
+  }
+
+  async uglyButton() {
+
+    let userId = (await this.afAuth.currentUser).uid;
+    // let uid = user;
+    // throw new Error('Method not implemented.');
+    // console.log(this.db.collection(`/users/${userId}/groups`).snapshotChanges().subscribe(snaps => {
+    //   console.log(snaps)
+    // }));
+    // await this.loadGroups().subscribe((res) => console.log("RES", res))
+    let test = await (this.loadGroups() as Observable<QuerySnapshot>);
+    console.log("this.loadGroups() as Observable<QuerySnapshot>)", test);
+  }
+  // let user = this.afAuth.auth.currentUser;
+  // uid = user.uid;
+
+  loadGroupsAndWords(): Observable<Group[]> {
+    return this.db.collection<Group>(`users/${this.userService.user.uid}/groups`).snapshotChanges()
       .pipe(
-        map((res: { message: string; data: any }) => {
-          return res.data.map((group): Group => {
-            return {
-              groupNum: group._id,
-              name: group.title,
-              averageProficiency: group.averageProficiency,
-            };
+        map((groups) => {
+          return groups.map((group) => {
+            const data = group.payload.doc.data();
+            const id = group.payload.doc.id;
+            const wordsRef = this.db.collection<Group>(`users/${this.userService.user.uid}/groups`).doc(id).collection<Word>('words');
+            const words$ = wordsRef.snapshotChanges().pipe(
+              map((words) =>
+                words.map((word) => {
+                  const wordData = word.payload.doc.data();
+                  const wordId = word.payload.doc.id;
+                  return { id: wordId, ...wordData };
+                })
+              )
+            );
+            return { id, ...data, words: words$ };
           });
         })
-      )
-      .subscribe((transformedPostData) => {
-        this.groups = transformedPostData;
-        this.groupsObs.next([...this.groups]);
-      });
+      );
   }
 
-  groupsObsListener() {
-    return this.groupsObs.asObservable();
+  deleteGroup(groupId: string) {
+    return from(this.db.doc(`users/${this.userService.user.uid}/groups/${groupId}`).delete());
   }
 
-  getAllGroups() {
-    return [...this.groups];
-  }
-  // Return single group, LATER ON SHOULD BE CHANGED TO ID SEARCH
-  //delete???? TODO
-  getGroup(name: string) {
-    const group: Group = this.groups.find((group) => {
-      return group.name === name;
-    });
-    return group;
+
+  deleteGroupAndWords(groupId: string)
+    : Observable<any> {
+    return this.db
+      .collection(`users/${this.userService.user.uid}/groups/${groupId}/words`)
+      .get()
+      .pipe(
+        concatMap((results) => {
+          const words = convertSnaps<Word>(results);
+          const batch = this.db.firestore.batch();
+          const groupRef = this.db.doc(`groups/${groupId}`).ref;
+          batch.delete(groupRef);
+          for (let word of words) {
+            const wordRef = this.db.doc(
+              `users/${this.userService.user.uid}/groups/${groupId}/words/${word.id}`
+            ).ref;
+            batch.delete(wordRef);
+          }
+          return from(batch.commit());
+        })
+      );
   }
 
-  editOrCreate(newName, groupNum?) {
-    if (groupNum) {
-      return this.http.put(this.BACKEND_URL + '/' + groupNum, {
-        title: newName,
-      });
-    } else {
-      return this.http.post(this.BACKEND_URL, {
-        title: newName,
-      });
-    }
-  }
-  deteleGroup(groupId) {
-    return this.http.delete(this.BACKEND_URL + '/' + groupId);
-  }
 }
