@@ -13,27 +13,45 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatRadioModule } from '@angular/material/radio';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TextFieldModule } from '@angular/cdk/text-field';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { NavbarComponent } from './navbar/navbar.component';
-import { HomepageComponent } from './homepage/homepage.component';
-import { MyVocabComponent } from './my-vocab/my-vocab.component';
-import { ExercisesComponent } from './exercises/exercises.component';
-import { SpellingComponent } from './exercises/spelling/spelling.component';
-import { ProgressBarComponent } from './exercises/progress-bar/progress-bar.component';
-import { ResultsComponent } from './exercises/results/results.component';
-import { RatingUIComponent } from './/my-vocab/rating-ui/rating-ui.component';
-import { WordFormComponent } from './my-vocab/word-form/word-form.component';
-import { GroupFormComponent } from './my-vocab/group-form/group-form.component';
-import { StopPropagationDirective } from './directives/stop-propagation.directive';
-import { AuthModule } from './auth/auth.module';
-import { AuthInterceptor } from './auth/auth-interceptor';
-import { AuthGuard } from './auth/auth.guard';
-import { SettingsPopupComponent } from './navbar/settings-popup/settings-popup.component';
-import { TypeTestComponent } from './exercises/type-test/type-test.component';
-import { ExtraTrCardComponent } from './my-vocab/extra-tr-card/extra-tr-card.component';
 import { HashLocationStrategy, LocationStrategy } from '@angular/common';
+
+// At App
+import { AppRoutingModule } from '@app/app-routing.module';
+import { AppComponent } from '@app/app.component';
+import { NavbarComponent } from '@app/navbar/navbar.component';
+import { HomepageComponent } from '@app/homepage/homepage.component';
+import { MyVocabComponent } from '@app/my-vocab/my-vocab.component';
+import { ExercisesComponent } from '@app/exercises/exercises.component';
+import { SpellingComponent } from '@app/exercises/spelling/spelling.component';
+import { ProgressBarComponent } from '@app/exercises/progress-bar/progress-bar.component';
+import { ResultsComponent } from '@app/exercises/results/results.component';
+import { RatingUIComponent } from '@app//my-vocab/rating-ui/rating-ui.component';
+import { WordFormComponent } from '@app/my-vocab/word-grid/word-form/word-form.component';
+import { GroupFormComponent } from '@app/my-vocab/group-form/group-form.component';
+import { StopPropagationDirective } from '@app/directives/stop-propagation.directive';
+import { SettingsPopupComponent } from '@app/navbar/settings-popup/settings-popup.component';
+import { TypeTestComponent } from '@app/exercises/type-test/type-test.component';
+import { WordGridComponent } from '@app/my-vocab/word-grid/word-grid.component';
+
+// Environment
+import { environment } from "@env/environment";
+
+// Firebase
+import { AngularFireModule } from "@angular/fire/compat";
+import {
+  AngularFirestoreModule,
+  USE_EMULATOR as USE_FIRESTORE_EMULATOR,
+} from "@angular/fire/compat/firestore";
+import {
+  AngularFireAuthModule,
+  USE_EMULATOR as USE_AUTH_EMULATOR,
+} from "@angular/fire/compat/auth";
+import {
+  AngularFireStorageModule,
+  USE_EMULATOR as USE_FUNCTIONS_EMULATOR,
+} from "@angular/fire/compat/storage";
+
+
 
 @NgModule({
   declarations: [
@@ -51,7 +69,6 @@ import { HashLocationStrategy, LocationStrategy } from '@angular/common';
     StopPropagationDirective,
     SettingsPopupComponent,
     TypeTestComponent,
-    ExtraTrCardComponent,
   ],
   imports: [
     BrowserModule,
@@ -68,15 +85,35 @@ import { HashLocationStrategy, LocationStrategy } from '@angular/common';
     BrowserAnimationsModule,
     ReactiveFormsModule,
     FormsModule,
-    AuthModule,
     MatIconModule,
     TextFieldModule,
+    AngularFireModule.initializeApp(environment.firebase.config),
+    AngularFireStorageModule,
+    AngularFireAuthModule,
+    AngularFirestoreModule,
+    WordGridComponent,
   ],
   providers: [
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
     { provide: LocationStrategy, useClass: HashLocationStrategy },
-    AuthGuard,
+    {
+      provide: USE_AUTH_EMULATOR,
+      useValue: environment.useEmulators
+        ? ["http://localhost:9099"]
+        : undefined,
+    },
+    {
+      provide: USE_FIRESTORE_EMULATOR,
+      useValue: environment.useEmulators
+        ? ["localhost", 8080]
+        : undefined,
+    },
+    {
+      provide: USE_FUNCTIONS_EMULATOR,
+      useValue: environment.useEmulators
+        ? ["http://localhost:5001"]
+        : undefined,
+    },
   ],
   bootstrap: [AppComponent],
 })
-export class AppModule {}
+export class AppModule { }
