@@ -1,6 +1,6 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { ProgressBarService } from 'src/app/services/progress-bar.service';
+import { ProgressBarService } from '@app/services/progress-bar.service';
 
 @Component({
   selector: 'app-progress-bar',
@@ -11,12 +11,12 @@ export class ProgressBarComponent implements OnInit, OnDestroy {
   private progressSub: Subscription;
   progress: number = -100;
 
-  constructor(private progressService: ProgressBarService) {}
+  constructor(private progressService: ProgressBarService) { }
 
   ngOnInit(): void {
     this.progressSub = this.progressService.progressEmitter.subscribe((res) => {
       this.progress += res;
     });
   }
-  ngOnDestroy(): void {}
+  ngOnDestroy(): void { }
 }
