@@ -8,7 +8,7 @@ import {
 } from '@angular/core';
 import { Word } from '../../models/word-model';
 import { ProgressBarService } from '../../services/progress-bar.service';
-import { WordManageService } from '../../services/word-manage.service';
+import { WordService } from '../../services/word.service';
 import { GroupService } from '../../services/group.service';
 import { SettingsService } from '../../services/settings.service';
 
@@ -47,14 +47,14 @@ export class SpellingComponent implements OnInit, AfterViewInit, OnDestroy {
   private wordSub: Subscription;
 
   constructor(
-    private wordService: WordManageService,
+    private wordService: WordService,
     private resultsService: ResultsService,
     private groupService: GroupService,
     private progressService: ProgressBarService,
     private settings: SettingsService,
     private router: Router,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   ngOnInit(): void {
     this.typeOfOS = window.navigator.platform;
@@ -64,18 +64,18 @@ export class SpellingComponent implements OnInit, AfterViewInit, OnDestroy {
     });
 
     // Subscribe to words
-    this.wordSub = this.wordService.wordsObsListener().subscribe(() => {
-      // Loading words
-      this.words = this.wordService.getRightWords(
-        this.exTypeParam,
-        this.groupIdParam
-      );
-      this.activeWord = this.words[this.activeWordIndex];
-      // Calculating percetage per word for progress bar
-      this.wordWorthPercent = 100 / this.words.length;
-      // Save array for results
-      this.resultsService.saveWordArr(this.words);
-    });
+    // this.wordSub = this.wordService.wordsObsListener().subscribe(() => {
+    //   // Loading words
+    //   this.words = this.wordService.getRightWords(
+    //     this.exTypeParam,
+    //     this.groupIdParam
+    //   );
+    //   this.activeWord = this.words[this.activeWordIndex];
+    //   // Calculating percetage per word for progress bar
+    //   this.wordWorthPercent = 100 / this.words.length;
+    //   // Save array for results
+    //   this.resultsService.saveWordArr(this.words);
+    // });
 
     // Settings config fetching
     this.translateDirection = this.settings.translateDirection;
@@ -91,7 +91,7 @@ export class SpellingComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    this.wordSub.unsubscribe();
+    // this.wordSub.unsubscribe();
   }
 
   // Handle correction and next word
@@ -127,13 +127,13 @@ export class SpellingComponent implements OnInit, AfterViewInit, OnDestroy {
     } else {
       // Break functionality when reached end of array
       if (this.activeWordIndex === this.words.length - 1) {
-        // TODO Redirect to RESULTS PAGE and send results
-        this.resultsService
-          .saveResultsInfo(this.scoreArr, 'spelling')
-          .subscribe(() => {
-            this.wordService.getWordsFromServer();
-            this.groupService.getGroupsFromServer();
-          });
+        // // TODO Redirect to RESULTS PAGE and send results
+        // this.resultsService
+        //   .saveResultsInfo(this.scoreArr, 'spelling')
+        //   .subscribe(() => {
+        //     this.wordService.getWordsFromServer();
+        //     this.groupService.loadGroups();
+        //   });
         this.router.navigate(['results'], { relativeTo: this.route });
         return;
       }
