@@ -9,11 +9,11 @@ import { environment } from '../../environments/environment';
 export class ResultsService {
   currentWordSet: Word[] = [];
   resultsArr: Boolean[] = [];
-  BACKEND_URL = environment.apiUrl;
-  formatedResults;
+  BACKEND_URL = "environment.apiUrl";
+  formattedResults;
   exMode: string;
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient) { }
 
   saveResultsInfo(resArr, exMode) {
     this.resultsArr = resArr;
@@ -21,15 +21,15 @@ export class ResultsService {
     this.formatResults();
     return this.http.put(
       this.BACKEND_URL + '/words/results',
-      this.formatedResults
+      this.formattedResults
     );
   }
 
-  // Formating results for server
+  // Formatting results for server
   formatResults() {
-    // We differentiate between modes to place cap on maximum score gained by using quizz
+    // We differentiate between modes to place cap on maximum score gained by using quiz
     if (this.exMode === 'quiz') {
-      return (this.formatedResults = this.currentWordSet
+      return (this.formattedResults = this.currentWordSet
         .map((word, index) => {
           let proficiency = 0;
           if (this.resultsArr[index] && word.proficiency < 14) proficiency = 1;
@@ -37,13 +37,11 @@ export class ResultsService {
           return {
             _id: word.id,
             proficiency: word.proficiency + proficiency,
-            group: word.groupNum,
-            user: word.user,
           };
         })
         .filter((result) => result !== undefined));
     } else if (this.exMode === 'spelling') {
-      return (this.formatedResults = this.currentWordSet
+      return (this.formattedResults = this.currentWordSet
         .map((word, index) => {
           let proficiency = -1;
           if (this.resultsArr[index] && word.proficiency < 20) proficiency = 1;
@@ -51,8 +49,6 @@ export class ResultsService {
           return {
             _id: word.id,
             proficiency: word.proficiency + proficiency,
-            group: word.groupNum,
-            user: word.user,
           };
         })
         .filter((result) => result !== undefined));
